[uwp] WIP 1 - Major overhaul of the DataGrid experience. See commit description for full details...
DetailRow
The DataGrid no longer has a selection mechanism. Now, each row has a DetailsRows that will let you see the contribution's specifics.

Cloning Old Contributions
you can CLONE any old contribution, update the details such as date and submit it as a new contribution. This is helpful for folks who re-use old activities because the specifics are the same as last year, but the date is different.

New Editor
Instead of having to go to a separate page to edit an item, a popup dialog appears instead,

Pause Refresh
This is a BIG one. Previous when navigating away or changing the data, this would trigger a fresh API call to get the latest list of contributions. However, this is a huge headache for folks who trigger a refresh after scrolling hundreds of items down and have to start at the top again.